DWCTerm object

R6 class to hold information on DWC terms

Format

R6Class object. Initialise a DwCTerm object

Methods

Public methods

DwCTerm$new()

DwCTerm$getQualifiedName()

DwCTerm$isDeprecated()

DwCTerm$print()

DwCTerm$getTermName()

DwCTerm$getNamespaceName()

DwCTerm$getTermIRI()

DwCTerm$getTermVersionIRI()

DwCTerm$getDateModified()

DwCTerm$getLabel()

DwCTerm$getIsReplacedBy()

DwCTerm$getDefinition()

DwCTerm$getNotes()

DwCTerm$getType()

DwCTerm$getExamples()

DwCTerm$getTermInformationLN()

DwCTerm$getExecCommitteeDecisions()

DwCTerm$getMiscInformation()

DwCTerm$getTermDef()

DwCTerm$getVocabularyURI()

DwCTerm$clone()

Method new()

Usage

DwCTerm$new(
  termName,
  namespaceName = character(),
  termIRI = character(),
  termVersionIRI = character(),
  dateModified = character(),
  label = character(),
  isReplacedBy = character(),
  definition = character(),
  notes = character(),
  type = character(),
  examples = character(),
  termInformationLN = character(),
  execCommitteeDecisions = character(),
  miscInformation = character(),
  termDef = character(),
  vocabularyURI = character()
)

Arguments

termName

A character scalar containing the name of the term

namespaceName

A character scalar containing the name of the namespace containing the term

termIRI

A character scalar containing the Internationalized Resource Identifier (IRI) of the definition of the term

termVersionIRI

A character scalar containing the Internationalized Resource Identifier (IRI) of the definition of the current version of the term

dateModified

A character scalar containing date information of the last modification of the term definition

label

A character scalar containing a label for the term (used in casual language)

isReplacedBy

A character scalar containing the Internationalized Resource Identifier (IRI) of the definition of the term that replaced this term (used if the term is deprecated)

definition

A character scalar containing a brief description of the purpose of the term

notes

A character vector containing extra information about the usage of the term

type

A character scalar containing the type of term defined

examples

A character vector giving a set of example values for the term

termInformationLN

A characer vector giving supplementary information for the term provided by the https://livingnorway.no/Living Norway initiative

execCommitteeDecisions

A character vector of links to decisions made by executive committees on the usage of the term

miscInformation

A character vector providing supplementary information for the term

termDef

A character scalar providing information about the location of the body in charge of definition of the term

vocabularyURI

A character scalar containing the Unified Resource Identifier (URI) for a vocabulary that the possible values for this term are based on

Returns

A new DwCTerm object Retrieve the qualified name of the term (the term name with the associated namespace)

Method getQualifiedName()

Usage

DwCTerm$getQualifiedName()

Returns

The qualified name of the term

Return the deprecation status of the term

Method isDeprecated()

Usage

DwCTerm$isDeprecated()

Returns

A logical scalar that is TRUE if the term is deprecated and FALSE otherwise

Print the term information

Retrieve the term name

Method print()

Usage

DwCTerm$print(...)

Method getTermName()

Usage

DwCTerm$getTermName()

Returns

A character scalar containing the term name

Retrieve the namespace name

Method getNamespaceName()

Usage

DwCTerm$getNamespaceName()

Returns

A character scalar containing the namespace name

Retrieve the term IRI

Method getTermIRI()

Usage

DwCTerm$getTermIRI()

Returns

A character scalar containing the term IRI

Retrieve the term version IRI

Method getTermVersionIRI()

Usage

DwCTerm$getTermVersionIRI()

Returns

A character scalar containing the term version IRI

Retrieve the modification date

Method getDateModified()

Usage

DwCTerm$getDateModified()

Returns

A character scalar containing the modification date

Retrieve the label

Method getLabel()

Usage

DwCTerm$getLabel()

Returns

A character scalar containing the label

Retrieve the replacement IRI

Method getIsReplacedBy()

Usage

DwCTerm$getIsReplacedBy()

Returns

A character scalar containing the IRI of the replacement term

Retrieve the definition

Method getDefinition()

Usage

DwCTerm$getDefinition()

Returns

A character scalar containing the term definition

Retrieve the term notes

Method getNotes()

Usage

DwCTerm$getNotes()

Returns

A character vector containing the term notes

Retrieve the type of the term

Method getType()

Usage

DwCTerm$getType()

Returns

A character scalar containing the term type Retrieve examples of the use of the term

Method getExamples()

Usage

DwCTerm$getExamples()

Returns

A character vector containing examples of use of the term

Retrieve supplementary information about the term provided by Living Norway

Method getTermInformationLN()

Usage

DwCTerm$getTermInformationLN()

Returns

A character vector containing the Living Norway supplementary information

Retrieve executive committee decisions about the use of the term

Method getExecCommitteeDecisions()

Usage

DwCTerm$getExecCommitteeDecisions()

Returns

A character vector containing the executive committee decisions

Retrieve miscellaneous information on the usage of the term

Method getMiscInformation()

Usage

DwCTerm$getMiscInformation()

Returns

A character vector containing the miscellaneous information

Retrieve the source of the term definition

Method getTermDef()

Usage

DwCTerm$getTermDef()

Returns

A character scalar containing the source of the term definition

Retrieve the vocabulary URI

Method getVocabularyURI()

Usage

DwCTerm$getVocabularyURI()

Returns

A character scalar containing the URI where the vocabulary for the term is described

Method clone()

The objects of this class are cloneable with this method.

Usage

DwCTerm$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.