R6 class representing a data structure for a GBIF Loan augmented data table (based on the Dawrin core)

R6 class representing a data structure for a GBIF Loan augmented data table (based on the Dawrin core)

R6 class representing a data structure for a GBIF Loan augmented data table (based on the Dawrin core)

Details

The GBIFLoan class allows for the specification of data tables that comply with the Loan http://data.ggbn.org/schemas/ggbn/terms/Loanclass specification of GBIF.

See also

Super class

LivingNorwayR::DwCGeneric -> GBIFLoan

Methods

Public methods

GBIFLoan$new()

GBIFLoan$clone()

Inherited methods

LivingNorwayR::DwCGeneric$exportAsDataFrame()

LivingNorwayR::DwCGeneric$exportTable()

LivingNorwayR::DwCGeneric$getAssociatedTerms()

LivingNorwayR::DwCGeneric$getDefaultYearFormat()

LivingNorwayR::DwCGeneric$getDwCClassName()

LivingNorwayR::DwCGeneric$getDwCTermInfo()

LivingNorwayR::DwCGeneric$getIDIndex()

LivingNorwayR::DwCGeneric$getIDName()

LivingNorwayR::DwCGeneric$getTableName()

LivingNorwayR::DwCGeneric$getTableTermName()

LivingNorwayR::DwCGeneric$getTermMapping()

LivingNorwayR::DwCGeneric$importDataTable()

LivingNorwayR::DwCGeneric$print()

LivingNorwayR::DwCGeneric$setDefaultYearFormat()

LivingNorwayR::DwCGeneric$setTableName()

Method new()

Create a new GBIFLoan object

Usage

GBIFLoan$new(
  objectData,
  idColumnInfo,
  nameAutoMap = FALSE,
  defDateFormat = "YYYY-MM-DD",
  ...
)

Arguments

objectData

A data.frame containing the data to import into the object

idColumnInfo

Either a character scalar containing the column name of objectData or an integer scalar giving the index of the column of objectData that corresponds to the ID variable. Alternatively, this parameter may be the qualified name of the Darwin core term for which the appropriately mapped column will be used as the ID variable (the possible Darwin core term names can be found by running names(getGBIFLoanMembers()))

nameAutoMap

A logical scalar that if TRUE maps the columns of objectData to their respective Darwin core terms based on the column names

defDateFormat

A character scalar providing the default format for strings denoting dates in the data table. See the https://dwc.tdwg.org/text/#1-introductionDarwin Core text guide for expected values for this string.

...

A named set of paramaeters corresponding to Darwin core terms associated with the GBIFLoan class type. Each is either a character scalar containing the column name of objectData or an integer scalar giving the index of the column of objectData that corresponds to the term. Mappable terms can be found using: sapply(X = getGBIFLoanMembers(), FUN = function(curTerm) { curTerm$getTermName() })

Returns

A new GBIFLoan object

Method clone()

The objects of this class are cloneable with this method.

Usage

GBIFLoan$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.