Initialise a metadata object from an import file

Initialise a metadata object from an import file

Usage

initializeDwCMetadata(fileLocation, fileEncoding = "", fileType = NA, ...)

Arguments

fileLocation

A character scalar containing the location of the import file

fileEncoding

A character string. If non-empty, declares the encoding to be used on a file so the character data can be re-encoded as they are written

fileType

A character scalar stating the type of the file. File type can be either "rmarkdown", "html", "eml", or "darwincore". If NA then the file type will be determined from the file extension

...

A set of arguments to be passed to the xml_new_root function to define namespace information. The values returned by the helper function getDefaultEMLDefinitionInfo will be used unless overwritten by user input here. In addition, arguments required by the EML standard but not provided here (such as the packageID attribute) will be otherwise auto-generated

Value

A new DwCMetadata object

See also