Initialize a new GBIF Identifier object

Initialize a new GBIF Identifier object

initializeGBIFIdentifier(
  objectData,
  idColumnInfo,
  nameAutoMap = FALSE,
  defDateFormat = "YYYY-MM-DD",
  ...
)

Arguments

objectData

A data.frame containing the data to import into the object

idColumnInfo

Either a character scalar containing the column name of objectData or an integer scalar giving the index of the column of objectData that corresponds to the ID variable. Alternatively, this parameter may be the qualified name of the Darwin core term for which the appropriately mapped column will be used as the ID variable (the possible Darwin core term names can be found by running names(getGBIFIdentifierMembers()))

nameAutoMap

A logical scalar that if TRUE maps the columns of objectData to their respective Darwin core terms based on the column names

defDateFormat

A character scalar providing the default format for strings denoting dates in the data table. See the https://dwc.tdwg.org/text/#1-introductionDarwin Core text guide for expected values for this string.

...

A named set of paramaeters corresponding to Darwin core terms associated with the GBIFIdentifier class type. Each is either a character scalar containing the column name of objectData or an integer scalar giving the index of the column of objectData that corresponds to the term. Mappable terms can be found using: sapply(X = getGBIFIdentifierMembers(), FUN = function(curTerm) { curTerm$getTermName() })

Value

A new GBIFIdentifier object

See also