function to search GBIFEventMembers

function to search GBIFEventMembers

searchGBIFEventMembers(term)

Arguments

term

the search term

Value

A subset of the list from the LivingNorwayR::getGBIFEventMembers() function

See also

Examples

searchGBIFEventMembers("georef")
#> http://rs.tdwg.org/dwc/terms/georeferencedBy - Georeferenced By
#> A list (concatenated and separated) of names of people, groups, or organizations who determined the georeference (spatial representation) for the Location.
#> 
#> 	Defined in: https://dwc.tdwg.org/
#> 	IRI: http://rs.tdwg.org/dwc/terms/georeferencedBy
#> 	Version IRI: http://rs.tdwg.org/dwc/terms/version/georeferencedBy-2017-10-06
#> 	Type: Property
#> 	Date modified: 2017-10-06
#> 	Notes:
#> 		Recommended best practice is to separate the values in a list with space vertical bar space ( | ).
#> 	Executive committee decisions:
#> 		http://rs.tdwg.org/decisions/decision-2014-10-30_16
#> 	Examples:
#> 		Brad Millen (ROM), Kristina Yamamoto | Janet Fang
#> 	Miscellaneous information:
#> 		not in ABCD
#> http://rs.tdwg.org/dwc/terms/georeferencedDate - Georeferenced Date
#> The date on which the Location was georeferenced.
#> 
#> 	Defined in: https://dwc.tdwg.org/
#> 	IRI: http://rs.tdwg.org/dwc/terms/georeferencedDate
#> 	Version IRI: http://rs.tdwg.org/dwc/terms/version/georeferencedDate-2020-08-12
#> 	Type: Property
#> 	Date modified: 2020-08-12
#> 	Notes:
#> 		Recommended best practice is to use a date that conforms to ISO 8601-1:2019.
#> 	Executive committee decisions:
#> 		http://rs.tdwg.org/decisions/decision-2011-10-16_9
#> 	Examples:
#> 		1963-03-08T14:07-0600 (8 Mar 1963 at 2:07pm in the time zone six hours earlier than UTC). 2009-02-20T08:40Z (20 February 2009 8:40am UTC). 2018-08-29T15:19 (3:19pm local time on 29 August 2018). 1809-02-12 (some time during 12 February 1809). 1906-06 (some time in June 1906). 1971 (some time in the year 1971). 2007-03-01T13:00:00Z/2008-05-11T15:30:00Z (some time during the interval between 1 March 2007 1pm UTC and 11 May 2008 3:30pm UTC). 1900/1909 (some time during the interval between the beginning of the year 1900 and the end of the year 1909). 2007-11-13/15 (some time in the interval between 13 November 2007 and 15 November 2007).
#> 	Miscellaneous information:
#> 		not in ABCD
#> http://rs.tdwg.org/dwc/terms/georeferenceProtocol - Georeference Protocol
#> A description or reference to the methods used to determine the spatial footprint, coordinates, and uncertainties.
#> 
#> 	Defined in: https://dwc.tdwg.org/
#> 	IRI: http://rs.tdwg.org/dwc/terms/georeferenceProtocol
#> 	Version IRI: http://rs.tdwg.org/dwc/terms/version/georeferenceProtocol-2020-08-20
#> 	Type: Property
#> 	Date modified: 2020-08-20
#> 	Examples:
#> 		Georeferencing Quick Reference Guide (Zermoglio et al. 2020, https://doi.org/10.35035/e09p-h128)
#> 	Miscellaneous information:
#> 		DataSets/DataSet/Units/Unit/Gathering/SiteCoordinateSets/SiteCoordinates/CoordinateMethod
#> http://rs.tdwg.org/dwc/terms/georeferenceSources - Georeference Sources
#> A list (concatenated and separated) of maps, gazetteers, or other resources used to georeference the Location, described specifically enough to allow anyone in the future to use the same resources.
#> 
#> 	Defined in: https://dwc.tdwg.org/
#> 	IRI: http://rs.tdwg.org/dwc/terms/georeferenceSources
#> 	Version IRI: http://rs.tdwg.org/dwc/terms/version/georeferenceSources-2020-10-28
#> 	Type: Property
#> 	Date modified: 2020-10-28
#> 	Notes:
#> 		Recommended best practice is to separate the values in a list with space vertical bar space ( | ).
#> 	Executive committee decisions:
#> 		http://rs.tdwg.org/decisions/decision-2014-10-30_16
#> 	Examples:
#> 		https://www.geonames.org/, USGS 1:24000 Florence Montana Quad 1967 | Terrametrics 2008 on Google Earth, GeoLocate
#> 	Miscellaneous information:
#> 		DataSets/DataSet/Units/Unit/Gathering/SiteCoordinateSets/SiteCoordinates/GeoreferenceSources
#> http://rs.tdwg.org/dwc/terms/georeferenceVerificationStatus - Georeference Verification Status
#> A categorical description of the extent to which the georeference has been verified to represent the best possible spatial description for the Location of the Occurrence.
#> 
#> 	Defined in: https://dwc.tdwg.org/
#> 	IRI: http://rs.tdwg.org/dwc/terms/georeferenceVerificationStatus
#> 	Version IRI: http://rs.tdwg.org/dwc/terms/version/georeferenceVerificationStatus-2021-07-15
#> 	Type: Property
#> 	Date modified: 2021-07-15
#> 	Notes:
#> 		Recommended best practice is to use a controlled vocabulary.
#> 	Executive committee decisions:
#> 		http://rs.tdwg.org/decisions/decision-2021-07-15_34
#> 	Examples:
#> 		unable to georeference, requires georeference, requires verification, verified by data custodian, verified by contributor
#> 	Miscellaneous information:
#> 		DataSets/DataSet/Units/Unit/Gathering/SiteCoordinateSets/SiteCoordinates/GeoreferenceVerificationStatus
#> http://rs.tdwg.org/dwc/terms/georeferenceRemarks - Georeference Remarks
#> Notes or comments about the spatial description determination, explaining assumptions made in addition or opposition to the those formalized in the method referred to in georeferenceProtocol.
#> 
#> 	Defined in: https://dwc.tdwg.org/
#> 	IRI: http://rs.tdwg.org/dwc/terms/georeferenceRemarks
#> 	Version IRI: http://rs.tdwg.org/dwc/terms/version/georeferenceRemarks-2017-10-06
#> 	Type: Property
#> 	Date modified: 2017-10-06
#> 	Examples:
#> 		Assumed distance by road (Hwy. 101).
#> 	Miscellaneous information:
#> 		DataSets/DataSet/Units/Unit/Gathering/SiteCoordinateSets/SiteCoordinates/GeoreferenceRemarks