R6 class that represents metadata in DwC

R6 class that represents metadata in DwC

R6 class that represents metadata in DwC

Format

R6Class object.

Details

The DwCMetadata class exists as a handling class for the encapsulation and handling of metadata being presented

Methods

Public methods

DwCMetadata$importFromLivingNorwayHTML()

DwCMetadata$importFromLivingNorwayRMD()

DwCMetadata$exportToEML()

DwCMetadata$new()

DwCMetadata$clone()

Method importFromLivingNorwayHTML()

Retrieve metadata information from a HTML file that has been created with a Living Norway HTML tag schema

Usage

DwCMetadata$importFromLivingNorwayHTML(fileLocation, fileEncoding = "")

Arguments

fileLocation

A character scalar containing the location of the HTML file

fileEncoding

A character string. If non-empty, declares the encoding to be used on a file so the character data can be re-encoded as they are written

Method importFromLivingNorwayRMD()

Retrieve metadata information from a RMD file that has been created with a Living Norway HTML tag schema

Usage

DwCMetadata$importFromLivingNorwayRMD(fileLocation, fileEncoding = "")

Arguments

fileLocation

A character scalar containing the location of the RMD file

fileEncoding

A character string. If non-empty, declares the encoding to be used on a file so the character data can be re-encoded as they are written

Method exportToEML()

Export the metadata as an EML XML file

Usage

DwCMetadata$exportToEML(fileLocation, fileEncoding = "")

Arguments

fileLocation

A character scalar containing the location to store the EML fiøe

fileEncoding

A character string. If non-empty, declares the encoding to be used on a file so the character data can be re-encoded as they are written

Method new()

Initialise a metadata object from an import file

Usage

DwCMetadata$new(fileLocation, fileEncoding = "", fileType = NA)

Arguments

fileLocation

A character scalar containing the location of the import file

fileEncoding

A character string. If non-empty, declares the encoding to be used on a file so the character data can be re-encoded as they are written

fileType

A character scalar stating the type of the file. File type can be either "rmarkdown", "html", "eml", or "darwincore". If NA then the file type will be determined from the file extension

Method clone()

The objects of this class are cloneable with this method.

Usage

DwCMetadata$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.