function to search GBIFOccurrenceMembers

function to search GBIFOccurrenceMembers

searchGBIFOccurenceMembers(term)

Arguments

term

the search term

Value

A subset of the list from the LivingNorwayR::getGBIFOccurrenceMembers() function

See also

Examples

searchGBIFOccurenceMembers("tax")
#> http://rs.tdwg.org/dwc/terms/taxonID - Taxon ID
#> An identifier for the set of taxon information (data associated with the Taxon class). May be a global unique identifier or an identifier specific to the data set.
#> 
#> 	Defined in: https://dwc.tdwg.org/
#> 	IRI: http://rs.tdwg.org/dwc/terms/taxonID
#> 	Version IRI: http://rs.tdwg.org/dwc/terms/version/taxonID-2017-10-06
#> 	Type: Property
#> 	Date modified: 2017-10-06
#> 	Examples:
#> 		8fa58e08-08de-4ac1-b69c-1235340b7001, 32567, https://www.gbif.org/species/212
#> 	Miscellaneous information:
#> 		not in ABCD
#> http://rs.tdwg.org/dwc/terms/taxonConceptID - Taxon Concept ID
#> An identifier for the taxonomic concept to which the record refers - not for the nomenclatural details of a taxon.
#> 
#> 	Defined in: https://dwc.tdwg.org/
#> 	IRI: http://rs.tdwg.org/dwc/terms/taxonConceptID
#> 	Version IRI: http://rs.tdwg.org/dwc/terms/version/taxonConceptID-2017-10-06
#> 	Type: Property
#> 	Date modified: 2017-10-06
#> 	Examples:
#> 		8fa58e08-08de-4ac1-b69c-1235340b7001
#> 	Miscellaneous information:
#> 		not in ABCD
#> http://rs.tdwg.org/dwc/terms/taxonRank - Taxon Rank
#> The taxonomic rank of the most specific name in the scientificName.
#> 
#> 	Defined in: https://dwc.tdwg.org/
#> 	IRI: http://rs.tdwg.org/dwc/terms/taxonRank
#> 	Version IRI: http://rs.tdwg.org/dwc/terms/version/taxonRank-2017-10-06
#> 	Type: Property
#> 	Date modified: 2017-10-06
#> 	Notes:
#> 		Recommended best practice is to use a controlled vocabulary.
#> 	Examples:
#> 		subspecies, varietas, forma, species, genus
#> 	Miscellaneous information:
#> 		DataSets/DataSet/Units/Unit/Identifications/Identification/TaxonIdentified/ScientificName/NameAtomised/Botanical/Rank
#> http://rs.tdwg.org/dwc/terms/taxonomicStatus - Taxonomic Status
#> The status of the use of the scientificName as a label for a taxon. Requires taxonomic opinion to define the scope of a taxon. Rules of priority then are used to define the taxonomic status of the nomenclature contained in that scope, combined with the experts opinion. It must be linked to a specific taxonomic reference that defines the concept.
#> 
#> 	Defined in: https://dwc.tdwg.org/
#> 	IRI: http://rs.tdwg.org/dwc/terms/taxonomicStatus
#> 	Version IRI: http://rs.tdwg.org/dwc/terms/version/taxonomicStatus-2017-10-06
#> 	Type: Property
#> 	Date modified: 2017-10-06
#> 	Notes:
#> 		Recommended best practice is to use a controlled vocabulary.
#> 	Examples:
#> 		invalid, misapplied, homotypic synonym, accepted
#> 	Miscellaneous information:
#> 		not in ABCD
#> http://rs.tdwg.org/dwc/terms/taxonRemarks - Taxon Remarks
#> Comments or notes about the taxon or name.
#> 
#> 	Defined in: https://dwc.tdwg.org/
#> 	IRI: http://rs.tdwg.org/dwc/terms/taxonRemarks
#> 	Version IRI: http://rs.tdwg.org/dwc/terms/version/taxonRemarks-2017-10-06
#> 	Type: Property
#> 	Date modified: 2017-10-06
#> 	Examples:
#> 		this name is a misspelling in common use
#> 	Miscellaneous information:
#> 		not in ABCD